+++ CHANGES.md
@@ -1,12 +1,19 @@
 # Targeting v0.7.1
 
 ## Important changes
+
+- `EvalMetrics` revised to inherit from `Callback` and be called on validation data after every epoch. User-written `EvalMetrics` willneed to be adjusted to work with the new calling method: adjust `evaluate` method and constructor may need to be adjusted; see existing metrics to see how.
   
 ## Breaking
+
+- `eval_metrics` argument in `train_models` renamed to `metric_partials` and now takes a list of partial `EvalMetrics`
+- User-written `EvalMetrics` will need to be adjusted to work with the new calling method: adjust `evaluate` method and constructor may need to be adjusted; see existing metrics to see how.
 
 ## Additions
 
 - `OneCycle` now has a `cycle_ends_training` which allows training to continue at the final LR and Momentum. keeping at default of `True` ends the training once the cycle is complete, as usual.
+- `to_np` now returns `None` when input tensor is `None`
+- `plot_train_history` now plots metric evolution for validation data
 
 ## Removals
 
@@ -14,10 +21,21 @@
 
 - `Model` now creates `cb_savepath` is it didn't already exist
 - Bug in `PredHandler` where predictions were kept on device leading to increased memory usage
+- Version issue in matplotlib affecting plot positioning
 
 ## Changes
 
 ## Depreciations
+
+- V0.8:
+    - All `EvalMetrics` depreciated with metric system. They have been copied and renamed to Old* for compatibility with the old model training system.
+    - `OldEvalMetric`: Replaced by `EvalMetric`
+    - `OldMultiAMS`: Replaced by `MultiAMS`
+    - `OldAMS`: Replaced by `AMS`
+    - `OldRegPull`: Replaced by `RegPull`
+    - `OldRegAsProxyPull`: Replaced by `RegAsProxyPull`
+    - `OldRocAucScore`: Replaced by `RocAucScore`
+    - `OldBinaryAccuracy`: Replaced by `BinaryAccuracy`
 
 ## Comments